refactor(debug): 优化调试环境配置和错误处理 fix #15, #16, #11
- 在 CppDebugEnv、JavaDebugEnv 和 PythonDebugEnv 中添加国际化支持- 改进用户提示信息，增加对环境变量设置的指导- 优化文件删除逻辑，提高代码健壮性
- 统一调试日志的输出格式- 修复了一些与调试环境相关的小问题

+++ src/main/java/com/xhf/leetcode/plugin/io/file/utils/FileUtils.java
@@ -304,7 +304,9 @@ public class FileUtils {
      * @throws IOException exp
      */
     public static void removeFile(String path) throws IOException {
-        Files.delete(Paths.get(path));
+        if (fileExists(path)) {
+            Files.delete(Paths.get(path));
+        }
     }
